+++ src/main/java/com/company/RssReaderBot/utils/Partition.java
@@ -1,4 +1,4 @@
-package com.company.RssReaderBot.services;
+package com.company.RssReaderBot.utils;
 
 import java.util.*;
