+++ src/main/java/com/company/RssReaderBot/utils/parser/RssParser.java
@@ -1,6 +1,7 @@
-package com.company.RssReaderBot.services.parser;
+package com.company.RssReaderBot.utils.parser;
 
-import com.company.RssReaderBot.entities.Item;
+import com.company.RssReaderBot.models.ItemModel;
+import com.company.RssReaderBot.utils.DateUtils;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
@@ -36,11 +37,17 @@ public class RssParser {
 
     @Getter @Setter
     private String rssUrl;
-
+    @Getter
     private String feedTitle;
+
+    private final DateUtils dateUtils;
 
     @Getter
     private NodeList nodeList;
+
+    public RssParser(DateUtils dateUtils) {
+        this.dateUtils = dateUtils;
+    }
 
     public void parseRss(String url) {
         try {
@@ -51,15 +58,13 @@ public class RssParser {
         }
     }
 
-    public String getFeedTitle(String url) {
+    public void parseFeedTitle(String url) {
         try {
             Document doc = getDocument(url);
             Element channel = (Element) doc.getElementsByTagName(CHANNEL).item(0);
             feedTitle = getOptionalSubElementValue(channel, TITLE);
-            return feedTitle;
         } catch (ParserConfigurationException | SAXException | IOException e) {
             e.printStackTrace();
-            return null;
         }
     }
 
@@ -70,66 +75,66 @@ public class RssParser {
         return doc;
     }
 
-    public List<Item> getAllElementsList() {
-        List<Item> itemsList = new ArrayList<>();
+    public List<ItemModel> getAllElementsList() {
+        List<ItemModel> itemsList = new ArrayList<>();
         for (int i = 0; i < nodeList.getLength(); i++) {
             Node node = nodeList.item(i);
             Element element = (Element) node;
             if (node.getNodeType() == Node.ELEMENT_NODE) {
-                Item currentItem = createItem(element);
-                itemsList.add(currentItem);
+                ItemModel currentItemModel = createItem(element);
+                itemsList.add(currentItemModel);
             }
         }
         return itemsList;
     }
 
-    public List<Item> getElementListByTitle(String title) {
-        List<Item> itemsList = new ArrayList<>();
+    public List<ItemModel> getElementListByTitle(String title) {
+        List<ItemModel> itemsList = new ArrayList<>();
         title = title.toLowerCase();
         for (int i = 0; i < nodeList.getLength(); i++) {
             Node node = nodeList.item(i);
             if (node.getNodeType() == Node.ELEMENT_NODE) {
                 Element element = (Element) node;
-                String currentNodeTitle = element.getElementsByTagName(TITLE).item(0).getTextContent().toLowerCase();
+                String currentNodeTitle = element
+                        .getElementsByTagName(TITLE)
+                        .item(0)
+                        .getTextContent()
+                        .toLowerCase();
                 if (currentNodeTitle.contains(title)) {
-                    Item currentItem = createItem(element);
-                    itemsList.add(currentItem);
+                    ItemModel currentItemModel = createItem(element);
+                    itemsList.add(currentItemModel);
                 }
             }
         }
         return itemsList;
     }
 
-    public List<Item> getElementListByDate(String dateString) {
-        List<Item> elementList = new ArrayList<>();
+    public List<ItemModel> getElementListByDate(String dateString) {
+        List<ItemModel> elementList = new ArrayList<>();
         for (int i = 0; i < nodeList.getLength(); i++) {
             Node node = nodeList.item(i);
             Element element = (Element) node;
             String currentItemDate = getOptionalSubElementValue(element, PUB_DATE);
             if (currentItemDate.contains(dateString)) {
-                Item currentItem = createItem(element);
-                elementList.add(currentItem);
+                ItemModel currentItemModel = createItem(element);
+                elementList.add(currentItemModel);
             }
         }
         return elementList;
     }
 
-    private Item createItem(Element element) {
+    private ItemModel createItem(Element element) {
         String itemTitle = getOptionalSubElementValue(element, TITLE);
         String itemDescription = getOptionalSubElementValue(element, DESCRIPTION);
-//        System.out.println(Optional.ofNullable(element.getElementsByTagName(GUID).item(0).getAttributes().getNamedItem("isPermaLink"))
-//                .map(Node::getTextContent).orElse(EMPTY_ELEMENT));
-        String pubDate = getOptionalSubElementValue(element, PUB_DATE);
+        Date pubDate = dateUtils.parseDate(getOptionalSubElementValue(element, PUB_DATE)).orElse(null);
         String guid = getOptionalSubElementValue(element, GUID);
-        // check enclosure (mediaUrl), if not null - insert, then go for link/guid, if null - do not use \n?
-        // check link first, then if not - check guid ispermalink=true
 
         String mediaUrl = Optional.ofNullable(element.getElementsByTagName(LINK).item(0))
                 .map(Node::getTextContent)
                 .orElseGet(() -> element.getElementsByTagName(ENCLOSURE).item(0)
                         .getAttributes().getNamedItem(URL).getTextContent());
 
-        return new Item(itemTitle, itemDescription, pubDate, mediaUrl, guid);
+        return new ItemModel(itemTitle, itemDescription, pubDate, mediaUrl, guid);
     }
 
     private String getOptionalSubElementValue(Element element, String tag) {
